// src/utils/types/ferramentas.ts
import { FabricanteI } from "./fabricante"

export interface Ferramenta {
  id: number
  modelo: string
  ano: number
  preco: number
  quantidadeEstoque: number
  destaque: boolean
  foto: string
  acessorios: string
  createdAt: Date
  updatedAt: Date
  tipo: string
  fabricante: FabricanteI
  fabricanteId: number
}